feat: separate client logic
