+++ src/app/match-driver/page.tsx
@@ -1,49 +1,9 @@
-'use client';
-
-import { useInfiniteQuery } from '@tanstack/react-query';
-import Link from 'next/link';
-import { getDriverListData } from '@/api/DriverService';
-import FindDriverCard from '@/components/cards/FindDriverCard';
 import SearchBar from '@/components/common/searchbar/SearchBar';
 import RegionServiceDropdown from '@/components/dropdown/RegionServiceDropdown';
 import SortDropdown from '@/components/dropdown/SortDropdown';
-import type { DriverListResponse } from '@/interfaces/API/DriverServiceInterface';
+import MatchDriverClient from '@/pages/MatchDriverClient';
 
 export default function MatchDriver() {
-  // const driverData = await getDriverListData(2);
-
-  // console.log(driverData);
-
-  const {
-    data: drivers,
-    isLoading: driversLoading,
-    error: driversError,
-    fetchNextPage,
-    hasNextPage,
-    isFetchingNextPage,
-  } = useInfiniteQuery<DriverListResponse>({
-    queryKey: ['drivers'],
-    queryFn: ({ pageParam }) => {
-      return getDriverListData(pageParam as number);
-    },
-    getNextPageParam: (lastPage, allPages) => {
-      const currentPage = allPages.length;
-      const totalPages = Math.ceil(lastPage.totalCount / 10);
-      return currentPage < totalPages ? currentPage + 1 : undefined;
-    },
-    initialPageParam: 1,
-  });
-
-  if (driversLoading) {
-    return <div>Loading...</div>;
-  }
-
-  if (driversError) {
-    return <div>Error</div>;
-  }
-
-  console.log(drivers);
-
   return (
     <div className="w-full flex items-center justify-center mb-[7rem]">
       <div className="lg:w-[120rem] sm:w-full items-start justify-center flex flex-col">
@@ -67,12 +27,7 @@ export default function MatchDriver() {
               </div>
             </div>
             <div className="w-full sm:px-[1rem] sm:gap-[2.4rem] md:gap-[3.2rem] lg:gap-[4.8rem] flex flex-col">
-              <FindDriverCard
-                data={drivers ? drivers.pages.flatMap(page => page.list) : []}
-                fetchNextPage={fetchNextPage}
-                hasNextPage={hasNextPage}
-                isFetchingNextPage={isFetchingNextPage}
-              />
+              <MatchDriverClient />
             </div>
           </div>
         </div>
